refactor(starVLA): consolidate HF evaluation scripts
Keep released-checkpoint launchers, preflights, checkpoint preparation, and reproducibility tests under policy/starVLA/scripts. Preserve the root setup_eval_* entry points required by XPolicyLab.
