+++ policy/starVLA/scripts/activate_policy_env.sh
