+++ policy/starVLA/scripts/prepare_hf_checkpoint.py
