+++ policy/starVLA/scripts/run_hf_robodojo_env_client.sh
