+++ policy/starVLA/scripts/run_robodojo_main.py
