+++ policy/starVLA/setup_eval_env_client.sh
@@ -29,7 +29,7 @@ yaml_file="${XPL_ROOT}/policy/${policy_name}/deploy.yml"
 echo "[CLIENT] policy=${policy_name}, task=${task_name}, ckpt_name=${ckpt_name}, server=${policy_server_host}:${policy_server_port}"
 
 if [[ "${STARVLA_DIRECT_ROBODOJO_CLIENT:-0}" == "1" ]]; then
-    exec bash "${SCRIPT_DIR}/tools/run_hf_robodojo_env_client.sh" \
+    exec bash "${SCRIPT_DIR}/scripts/run_hf_robodojo_env_client.sh" \
         "${eval_env_conda_env}" \
         "${policy_server_port}" \
         "${bench_name}" \
