--- policy/starVLA/tools/__init__.py
@@ -1,1 +1,0 @@
-"""Utilities for the XPolicyLab StarVLA adapter."""
